Fixed the node count on the discovery page
I've also merged some code that was shared between the dashboard and the
setup controllers.

Fixes bsc#1036040
